+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Changed
 
+- **Breaking:** Remove CUTLASS dependency and `wp.matmul()` functionality (including batched version).
 - `warp.fem.integrate()` and `warp.fem.integrate()` may now perform parallel evaluation of quadrature points within elements
 - `warp.fem.interpolate()` can now build Jacobian sparse matrices of interpolated functions with respect to a trial field